requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.19
django-adaptors==0.2.5
django-allauth==0.54.0
django-celery==3.1.17
django-filter==23.2
django-js-asset==2.1.0
django-mptt==0.14.0
django-rest-auth==0.9.5
django-tables2==2.5.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
kombu==5.2.4
load-dotenv==0.1.0
lxml==4.9.3
names==0.3.0
oauthlib==3.2.2
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
pydub==0.25.1
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.2
uuid==1.30
vine==5.0.0
wcwidth==0.2.6
